dual sparse table

+++ verify/AOJ/datastructure/dual_sparse_table.test.cpp
@@ -5,13 +5,13 @@
 int main(){
   int n,q;
   std::cin>>n>>q;
-  std::vector<std::tuple<int,int,std::pair<int,int>>>query(q);
+  std::vector<std::tuple<int,int,std::pair<int,long long>>>query(q);
   for(int i=0;i<q;i++){
     int l,k;
     std::cin>>l>>k;
     l--;
     query[i]={l,l+k,{1,1-l}};
   }
-  auto ans=dual_sparse_table<MonoidAddPair<int,int>>(n,query);
+  auto ans=dual_sparse_table<MonoidAddPair<int,long long>>(n,query);
   for(int i=0;i<n;i++)std::cout<<(long long)ans[i].first*i+ans[i].second<<" \n"[i+1==n];
 }
